chore: extracting state from uploader and creating a hook
